Move unloaded structs to dedicated file
